refactor: extract common type to shared/types
