+++ src/clients/Governance.ts
@@ -4,7 +4,6 @@ import env from 'decentraland-gatsby/dist/utils/env'
 import snakeCase from 'lodash/snakeCase'
 
 import { AirdropOutcome } from '../back/types/AirdropJob'
-import { NewsletterSubscriptionResult } from '../back/types/newsletter'
 import { SpecState } from '../components/Debug/UploadBadgeSpec'
 import { GOVERNANCE_API } from '../constants'
 import { BadgeCreationResult, RevokeOrReinstateResult, UserBadges } from '../entities/Badges/types'
@@ -36,6 +35,7 @@ import { SubscriptionAttributes } from '../entities/Subscription/types'
 import { Topic } from '../entities/SurveyTopic/types'
 import { ProjectHealth, UpdateAttributes, UpdateResponse } from '../entities/Updates/types'
 import { Vote, VotedProposal, Voter } from '../entities/Votes/types'
+import { NewsletterSubscriptionResult } from '../shared/types/newsletter'
 import Time from '../utils/date/Time'
 
 import { TransparencyBudget } from './DclData'
