+++ src/components/Modal/NewsletterSubscriptionModal/NewsletterSubscriptionModal.tsx
@@ -9,10 +9,10 @@ import { Header } from 'decentraland-ui/dist/components/Header/Header'
 import { Modal, ModalProps } from 'decentraland-ui/dist/components/Modal/Modal'
 import isEmail from 'validator/lib/isEmail'
 
-import { NewsletterSubscriptionResult } from '../../../back/types/newsletter'
 import { Governance } from '../../../clients/Governance'
 import useAsyncTask from '../../../hooks/useAsyncTask'
 import useFormatMessage from '../../../hooks/useFormatMessage'
+import { NewsletterSubscriptionResult } from '../../../shared/types/newsletter'
 import { ANONYMOUS_USR_SUBSCRIPTION, NEWSLETTER_SUBSCRIPTION_KEY } from '../../Banner/Subscription/SubscriptionBanner'
 import Label from '../../Common/Typography/Label'
 import Text from '../../Common/Typography/Text'
