+++ src/shared/types/newsletter.ts
